Added major version check and confirmation step, in regards to PR suggestions.

+++ packages/wrangler/src/api/upgrade.ts
@@ -3,12 +3,18 @@ import path from "path";
 import { execaSync } from "execa";
 import { fetch } from "undici";
 import { findWranglerToml } from "../config";
+import { confirm } from "../dialogs";
 import { printWranglerBanner } from "../index";
 import { logger } from "../logger";
 import { readFileSync } from "../parse";
+
 import type { PackageJSON } from "../parse";
 
-export async function upgradeWrangler() {
+export async function upgradeWrangler({
+	yesFlag = false,
+}: {
+	yesFlag: boolean;
+}) {
 	await printWranglerBanner();
 	let packageJsonData: undefined | PackageJSON;
 
@@ -29,6 +35,24 @@ export async function upgradeWrangler() {
 					}
 				)["dist-tags"];
 				const latestVersion = wranglerDistTags.latest;
+
+				const currentVersion = packageJsonData.devDependencies
+					.wrangler as string;
+				logger.log(
+					`Attempting to upgrade Wrangler from ${currentVersion} to the latest version ${latestVersion}...`
+				);
+
+				if (
+					latestVersion.split(".")[0] !== currentVersion.split(".")[0] &&
+					!yesFlag
+				) {
+					const userDecision = await confirm(
+						`⚠️  A major semver change has been detected. Would you like to continue?`
+					);
+					if (userDecision === false) {
+						return;
+					}
+				}
 
 				packageJsonData.devDependencies.wrangler = latestVersion;
 				writeFileSync(
